feat(all): Refactored props handling, removed typestyle
PropsNormalizer was renamed to ClassesExtractor and only extracts class object and classNames. The
provided props left intact. The style object is no longer tranformed to a classname via typestyle
and is left inside the provided props. It is added to the snabbdom data object via the style module.
Common used methods were moved to SMCComponent and SMCAdapter and inheriting classes can override
methods to customize behavior.

BREAKING CHANGE: style object is no removed from props and added to class via typestyle. In turn it
is available inside the snabbdom data object.